+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/).
 
 ## [Unreleased]
 
+## [1.16.1] - 2026-07-07
+
 ### Fixed
 
 - `Helpers::fieldFormatter()`: a surfaced ACF field object with no saved value